Move editor docks to a dedicated singleton
Remove docks getter from EditorInterface.
Add EditorDocks getter to EditorPlugin
