+++ editor/editor_audio_buses.cpp
@@ -1176,7 +1176,7 @@ void EditorAudioBuses::_server_save() {
 }
 
 void EditorAudioBuses::_select_layout() {
-	EditorNode::get_singleton()->get_filesystem_dock()->select_file(edited_path);
+	EditorDocks::get_singleton()->get_filesystem_dock()->select_file(edited_path);
 }
 
 void EditorAudioBuses::_save_as_layout() {
