+++ editor/editor_node.cpp
@@ -414,7 +414,7 @@ void EditorNode::_unhandled_input(const Ref<InputEvent> &p_event) {
 			_scene_tab_changed(next_tab);
 		}
 		if (ED_IS_SHORTCUT("editor/filter_files", p_event)) {
-			filesystem_dock->focus_on_filter();
+			EditorDocks::get_singleton()->get_filesystem_dock()->focus_on_filter();
 		}
 
 		if (ED_IS_SHORTCUT("editor/editor_2d", p_event)) {
@@ -1069,7 +1069,7 @@ Error EditorNode::load_resource(const String &p_resource, bool p_ignore_broken_d
 		return ERR_FILE_MISSING_DEPENDENCIES;
 	}
 
-	inspector_dock->edit_resource(res);
+	EditorDocks::get_singleton()->get_inspector_dock()->edit_resource(res);
 	return OK;
 }
 
@@ -1970,9 +1970,9 @@ void EditorNode::edit_item(Object *p_object) {
 void EditorNode::push_item(Object *p_object, const String &p_property, bool p_inspector_only) {
 	if (!p_object) {
 		get_inspector()->edit(nullptr);
-		node_dock->set_node(nullptr);
-		scene_tree_dock->set_selected(nullptr);
-		inspector_dock->update(nullptr);
+		EditorDocks::get_singleton()->get_node_dock()->set_node(nullptr);
+		EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(nullptr);
+		EditorDocks::get_singleton()->get_inspector_dock()->update(nullptr);
 		_display_top_editors(false);
 		return;
 	}
@@ -2037,10 +2037,10 @@ void EditorNode::_edit_current() {
 	this->current = current_obj;
 
 	if (!current_obj) {
-		scene_tree_dock->set_selected(nullptr);
+		EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(nullptr);
 		get_inspector()->edit(nullptr);
-		node_dock->set_node(nullptr);
-		inspector_dock->update(nullptr);
+		EditorDocks::get_singleton()->get_node_dock()->set_node(nullptr);
+		EditorDocks::get_singleton()->get_inspector_dock()->update(nullptr);
 
 		_display_top_editors(false);
 
@@ -2060,10 +2060,10 @@ void EditorNode::_edit_current() {
 		Resource *current_res = Object::cast_to<Resource>(current_obj);
 		ERR_FAIL_COND(!current_res);
 		get_inspector()->edit(current_res);
-		scene_tree_dock->set_selected(nullptr);
-		node_dock->set_node(nullptr);
-		inspector_dock->update(nullptr);
-		EditorNode::get_singleton()->get_import_dock()->set_edit_path(current_res->get_path());
+		EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(nullptr);
+		EditorDocks::get_singleton()->get_node_dock()->set_node(nullptr);
+		EditorDocks::get_singleton()->get_inspector_dock()->update(nullptr);
+		EditorDocks::get_singleton()->get_import_dock()->set_edit_path(current_res->get_path());
 
 		int subr_idx = current_res->get_path().find("::");
 		if (subr_idx != -1) {
@@ -2086,13 +2086,13 @@ void EditorNode::_edit_current() {
 
 		get_inspector()->edit(current_node);
 		if (current_node->is_inside_tree()) {
-			node_dock->set_node(current_node);
-			scene_tree_dock->set_selected(current_node);
-			inspector_dock->update(current_node);
+			EditorDocks::get_singleton()->get_node_dock()->set_node(current_node);
+			EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(current_node);
+			EditorDocks::get_singleton()->get_inspector_dock()->update(current_node);
 		} else {
-			node_dock->set_node(nullptr);
-			scene_tree_dock->set_selected(nullptr);
-			inspector_dock->update(nullptr);
+			EditorDocks::get_singleton()->get_node_dock()->set_node(nullptr);
+			EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(nullptr);
+			EditorDocks::get_singleton()->get_inspector_dock()->update(nullptr);
 		}
 
 		if (get_edited_scene() && get_edited_scene()->get_filename() != String()) {
@@ -2132,9 +2132,9 @@ void EditorNode::_edit_current() {
 		}
 
 		get_inspector()->edit(current_obj);
-		node_dock->set_node(nullptr);
-		scene_tree_dock->set_selected(selected_node);
-		inspector_dock->update(nullptr);
+		EditorDocks::get_singleton()->get_node_dock()->set_node(nullptr);
+		EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(selected_node);
+		EditorDocks::get_singleton()->get_inspector_dock()->update(nullptr);
 	}
 
 	if (current_obj == prev_inspected_object) {
@@ -2142,7 +2142,7 @@ void EditorNode::_edit_current() {
 		get_inspector()->update_tree();
 	}
 
-	inspector_dock->set_warning(editable_warning);
+	EditorDocks::get_singleton()->get_inspector_dock()->set_warning(editable_warning);
 
 	if (get_inspector()->is_capitalize_paths_enabled() != capitalize) {
 		get_inspector()->set_enable_capitalize_paths(capitalize);
@@ -2215,8 +2215,8 @@ void EditorNode::_edit_current() {
 		}
 	}
 
-	inspector_dock->update(current_obj);
-	inspector_dock->update_keying();
+	EditorDocks::get_singleton()->get_inspector_dock()->update(current_obj);
+	EditorDocks::get_singleton()->get_inspector_dock()->update_keying();
 }
 
 void EditorNode::_run(bool p_current, const String &p_custom) {
@@ -2669,7 +2669,7 @@ void EditorNode::_menu_option_confirm(int p_option, bool p_confirmed) {
 		case FILE_SHOW_IN_FILESYSTEM: {
 			String path = editor_data.get_scene_path(editor_data.get_edited_scene());
 			if (path != String()) {
-				filesystem_dock->navigate_to_path(path);
+				EditorDocks::get_singleton()->get_filesystem_dock()->navigate_to_path(path);
 			}
 		} break;
 
@@ -3252,7 +3252,7 @@ void EditorNode::set_edited_scene(Node *p_scene) {
 	if (Object::cast_to<Popup>(p_scene)) {
 		Object::cast_to<Popup>(p_scene)->show(); //show popups
 	}
-	scene_tree_dock->set_edited_scene(p_scene);
+	EditorDocks::get_singleton()->get_scene_tree_dock()->set_edited_scene(p_scene);
 	if (get_tree()) {
 		get_tree()->set_edited_scene_root(p_scene);
 	}
@@ -3277,10 +3277,10 @@ int EditorNode::_get_current_main_editor() {
 Dictionary EditorNode::_get_main_scene_state() {
 	Dictionary state;
 	state["main_tab"] = _get_current_main_editor();
-	state["scene_tree_offset"] = scene_tree_dock->get_tree_editor()->get_scene_tree()->get_vscroll_bar()->get_value();
+	state["scene_tree_offset"] = EditorDocks::get_singleton()->get_scene_tree_dock()->get_tree_editor()->get_scene_tree()->get_vscroll_bar()->get_value();
 	state["property_edit_offset"] = get_inspector()->get_scroll_offset();
 	state["saved_version"] = saved_version;
-	state["node_filter"] = scene_tree_dock->get_filter();
+	state["node_filter"] = EditorDocks::get_singleton()->get_scene_tree_dock()->get_filter();
 	return state;
 }
 
@@ -3322,14 +3322,14 @@ void EditorNode::_set_main_scene_state(Dictionary p_state, Node *p_for_scene) {
 	}
 
 	if (p_state.has("scene_tree_offset")) {
-		scene_tree_dock->get_tree_editor()->get_scene_tree()->get_vscroll_bar()->set_value(p_state["scene_tree_offset"]);
+		EditorDocks::get_singleton()->get_scene_tree_dock()->get_tree_editor()->get_scene_tree()->get_vscroll_bar()->set_value(p_state["scene_tree_offset"]);
 	}
 	if (p_state.has("property_edit_offset")) {
 		get_inspector()->set_scroll_offset(p_state["property_edit_offset"]);
 	}
 
 	if (p_state.has("node_filter")) {
-		scene_tree_dock->set_filter(p_state["node_filter"]);
+		EditorDocks::get_singleton()->get_scene_tree_dock()->set_filter(p_state["node_filter"]);
 	}
 
 	//this should only happen at the very end
@@ -3384,7 +3384,7 @@ void EditorNode::set_current_scene(int p_idx) {
 		Object::cast_to<Popup>(new_scene)->show(); //show popups
 	}
 
-	scene_tree_dock->set_edited_scene(new_scene);
+	EditorDocks::get_singleton()->get_scene_tree_dock()->set_edited_scene(new_scene);
 	if (get_tree()) {
 		get_tree()->set_edited_scene_root(new_scene);
 	}
@@ -3564,7 +3564,7 @@ Error EditorNode::load_scene(const String &p_scene, bool p_ignore_broken_deps, b
 
 	prev_scene->set_disabled(previous_scenes.size() == 0);
 	opening_prev = false;
-	scene_tree_dock->set_selected(new_scene);
+	EditorDocks::get_singleton()->get_scene_tree_dock()->set_selected(new_scene);
 
 	EditorDebuggerNode::get_singleton()->update_live_edit_root();
 
@@ -3589,27 +3589,11 @@ void EditorNode::open_request(const String &p_path) {
 }
 
 void EditorNode::request_instance_scene(const String &p_path) {
-	scene_tree_dock->instance(p_path);
+	EditorDocks::get_singleton()->get_scene_tree_dock()->instance(p_path);
 }
 
 void EditorNode::request_instance_scenes(const Vector<String> &p_files) {
-	scene_tree_dock->instance_scenes(p_files);
-}
-
-ImportDock *EditorNode::get_import_dock() {
-	return import_dock;
-}
-
-FileSystemDock *EditorNode::get_filesystem_dock() {
-	return filesystem_dock;
-}
-
-SceneTreeDock *EditorNode::get_scene_tree_dock() {
-	return scene_tree_dock;
-}
-
-InspectorDock *EditorNode::get_inspector_dock() {
-	return inspector_dock;
+	EditorDocks::get_singleton()->get_scene_tree_dock()->instance_scenes(p_files);
 }
 
 void EditorNode::_inherit_request(String p_file) {
@@ -3763,6 +3747,7 @@ void EditorNode::register_editor_types() {
 	ClassDB::register_virtual_class<ScriptEditorBase>();
 	ClassDB::register_class<EditorSyntaxHighlighter>();
 	ClassDB::register_virtual_class<EditorInterface>();
+	ClassDB::register_virtual_class<EditorDocks>();
 	ClassDB::register_class<EditorExportPlugin>();
 	ClassDB::register_class<EditorResourceConversionPlugin>();
 	ClassDB::register_class<EditorSceneImporter>();
@@ -4383,10 +4368,10 @@ void EditorNode::_save_docks_to_config(Ref<ConfigFile> p_layout, const String &p
 		}
 	}
 
-	p_layout->set_value(p_section, "dock_filesystem_split", filesystem_dock->get_split_offset());
-	p_layout->set_value(p_section, "dock_filesystem_display_mode", filesystem_dock->get_display_mode());
-	p_layout->set_value(p_section, "dock_filesystem_file_sort", filesystem_dock->get_file_sort());
-	p_layout->set_value(p_section, "dock_filesystem_file_list_display_mode", filesystem_dock->get_file_list_display_mode());
+	p_layout->set_value(p_section, "dock_filesystem_split", EditorDocks::get_singleton()->get_filesystem_dock()->get_split_offset());
+	p_layout->set_value(p_section, "dock_filesystem_display_mode", EditorDocks::get_singleton()->get_filesystem_dock()->get_display_mode());
+	p_layout->set_value(p_section, "dock_filesystem_file_sort", EditorDocks::get_singleton()->get_filesystem_dock()->get_file_sort());
+	p_layout->set_value(p_section, "dock_filesystem_file_list_display_mode", EditorDocks::get_singleton()->get_filesystem_dock()->get_file_list_display_mode());
 
 	for (int i = 0; i < vsplits.size(); i++) {
 		if (vsplits[i]->is_visible_in_tree()) {
@@ -4572,22 +4557,22 @@ void EditorNode::_load_docks_from_config(Ref<ConfigFile> p_layout, const String 
 
 	if (p_layout->has_section_key(p_section, "dock_filesystem_split")) {
 		int fs_split_ofs = p_layout->get_value(p_section, "dock_filesystem_split");
-		filesystem_dock->set_split_offset(fs_split_ofs);
+		EditorDocks::get_singleton()->get_filesystem_dock()->set_split_offset(fs_split_ofs);
 	}
 
 	if (p_layout->has_section_key(p_section, "dock_filesystem_display_mode")) {
 		FileSystemDock::DisplayMode dock_filesystem_display_mode = FileSystemDock::DisplayMode(int(p_layout->get_value(p_section, "dock_filesystem_display_mode")));
-		filesystem_dock->set_display_mode(dock_filesystem_display_mode);
+		EditorDocks::get_singleton()->get_filesystem_dock()->set_display_mode(dock_filesystem_display_mode);
 	}
 
 	if (p_layout->has_section_key(p_section, "dock_filesystem_file_sort")) {
 		FileSystemDock::FileSortOption dock_filesystem_file_sort = FileSystemDock::FileSortOption(int(p_layout->get_value(p_section, "dock_filesystem_file_sort")));
-		filesystem_dock->set_file_sort(dock_filesystem_file_sort);
+		EditorDocks::get_singleton()->get_filesystem_dock()->set_file_sort(dock_filesystem_file_sort);
 	}
 
 	if (p_layout->has_section_key(p_section, "dock_filesystem_file_list_display_mode")) {
 		FileSystemDock::FileListDisplayMode dock_filesystem_file_list_display_mode = FileSystemDock::FileListDisplayMode(int(p_layout->get_value(p_section, "dock_filesystem_file_list_display_mode")));
-		filesystem_dock->set_file_list_display_mode(dock_filesystem_file_list_display_mode);
+		EditorDocks::get_singleton()->get_filesystem_dock()->set_file_list_display_mode(dock_filesystem_file_list_display_mode);
 	}
 
 	for (int i = 0; i < vsplits.size(); i++) {
@@ -4805,7 +4790,7 @@ void EditorNode::_layout_menu_option(int p_id) {
 void EditorNode::_scene_tab_script_edited(int p_tab) {
 	Ref<Script> script = editor_data.get_scene_root_script(p_tab);
 	if (script.is_valid()) {
-		inspector_dock->edit_resource(script);
+		EditorDocks::get_singleton()->get_inspector_dock()->edit_resource(script);
 	}
 }
 
@@ -5260,7 +5245,7 @@ void EditorNode::_global_menu_new_window(const Variant &p_tag) {
 }
 
 void EditorNode::_dropped_files(const Vector<String> &p_files, int p_screen) {
-	String to_path = ProjectSettings::get_singleton()->globalize_path(get_filesystem_dock()->get_selected_path());
+	String to_path = ProjectSettings::get_singleton()->globalize_path(EditorDocks::get_singleton()->get_filesystem_dock()->get_selected_path());
 
 	_add_dropped_files_recursive(p_files, to_path);
 
@@ -5488,15 +5473,15 @@ void EditorNode::_resource_loaded(RES p_resource, const String &p_path) {
 
 void EditorNode::_feature_profile_changed() {
 	Ref<EditorFeatureProfile> profile = feature_profile_manager->get_current_profile();
-	TabContainer *import_tabs = cast_to<TabContainer>(import_dock->get_parent());
-	TabContainer *node_tabs = cast_to<TabContainer>(node_dock->get_parent());
-	TabContainer *fs_tabs = cast_to<TabContainer>(filesystem_dock->get_parent());
+	TabContainer *import_tabs = cast_to<TabContainer>(EditorDocks::get_singleton()->get_import_dock()->get_parent());
+	TabContainer *node_tabs = cast_to<TabContainer>(EditorDocks::get_singleton()->get_node_dock()->get_parent());
+	TabContainer *fs_tabs = cast_to<TabContainer>(EditorDocks::get_singleton()->get_filesystem_dock()->get_parent());
 	if (profile.is_valid()) {
-		node_tabs->set_tab_hidden(node_dock->get_index(), profile->is_feature_disabled(EditorFeatureProfile::FEATURE_NODE_DOCK));
+		node_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_node_dock()->get_index(), profile->is_feature_disabled(EditorFeatureProfile::FEATURE_NODE_DOCK));
 		// The Import dock is useless without the FileSystem dock. Ensure the configuration is valid.
 		bool fs_dock_disabled = profile->is_feature_disabled(EditorFeatureProfile::FEATURE_FILESYSTEM_DOCK);
-		fs_tabs->set_tab_hidden(filesystem_dock->get_index(), fs_dock_disabled);
-		import_tabs->set_tab_hidden(import_dock->get_index(), fs_dock_disabled || profile->is_feature_disabled(EditorFeatureProfile::FEATURE_IMPORT_DOCK));
+		fs_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_filesystem_dock()->get_index(), fs_dock_disabled);
+		import_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_import_dock()->get_index(), fs_dock_disabled || profile->is_feature_disabled(EditorFeatureProfile::FEATURE_IMPORT_DOCK));
 
 		main_editor_buttons[EDITOR_3D]->set_visible(!profile->is_feature_disabled(EditorFeatureProfile::FEATURE_3D));
 		main_editor_buttons[EDITOR_SCRIPT]->set_visible(!profile->is_feature_disabled(EditorFeatureProfile::FEATURE_SCRIPT));
@@ -5509,12 +5494,12 @@ void EditorNode::_feature_profile_changed() {
 			_editor_select(EDITOR_2D);
 		}
 	} else {
-		import_tabs->set_tab_hidden(import_dock->get_index(), false);
-		node_tabs->set_tab_hidden(node_dock->get_index(), false);
-		fs_tabs->set_tab_hidden(filesystem_dock->get_index(), false);
-		import_dock->set_visible(true);
-		node_dock->set_visible(true);
-		filesystem_dock->set_visible(true);
+		import_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_import_dock()->get_index(), false);
+		node_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_node_dock()->get_index(), false);
+		fs_tabs->set_tab_hidden(EditorDocks::get_singleton()->get_filesystem_dock()->get_index(), false);
+		EditorDocks::get_singleton()->get_import_dock()->set_visible(true);
+		EditorDocks::get_singleton()->get_node_dock()->set_visible(true);
+		EditorDocks::get_singleton()->get_filesystem_dock()->set_visible(true);
 		main_editor_buttons[EDITOR_3D]->set_visible(true);
 		main_editor_buttons[EDITOR_SCRIPT]->set_visible(true);
 		if (StreamPeerSSL::is_available()) {
@@ -6536,12 +6521,12 @@ EditorNode::EditorNode() {
 
 	// Instantiate and place editor docks
 
-	scene_tree_dock = memnew(SceneTreeDock(this, scene_root, editor_selection, editor_data));
-	inspector_dock = memnew(InspectorDock(this, editor_data));
-	import_dock = memnew(ImportDock);
-	node_dock = memnew(NodeDock);
+	SceneTreeDock *scene_tree_dock = memnew(SceneTreeDock(this, scene_root, editor_selection, editor_data));
+	InspectorDock *inspector_dock = memnew(InspectorDock(this, editor_data));
+	ImportDock *import_dock = memnew(ImportDock);
+	NodeDock *node_dock = memnew(NodeDock);
 
-	filesystem_dock = memnew(FileSystemDock(this));
+	FileSystemDock *filesystem_dock = memnew(FileSystemDock(this));
 	filesystem_dock->connect("inherit", callable_mp(this, &EditorNode::_inherit_request));
 	filesystem_dock->connect("instance", callable_mp(this, &EditorNode::_instance_request));
 	filesystem_dock->connect("display_mode_changed", callable_mp(this, &EditorNode::_save_docks));
@@ -6742,6 +6727,20 @@ EditorNode::EditorNode() {
 
 	preview_gen = memnew(AudioStreamPreviewGenerator);
 	add_child(preview_gen);
+
+	// Initialize singletons before adding plugins
+
+	EditorInterface *editor_interface = memnew(EditorInterface);
+	add_child(editor_interface);
+	EditorDocks *editor_docks = memnew(EditorDocks);
+	add_child(editor_docks);
+	editor_docks->scene_tree_dock = scene_tree_dock;
+	editor_docks->filesystem_dock = filesystem_dock;
+	editor_docks->node_dock = node_dock;
+	//	editor_docks->connection_dock = node_dock->connections;
+	editor_docks->inspector_dock = inspector_dock;
+	editor_docks->import_dock = import_dock;
+
 	//plugin stuff
 
 	add_editor_plugin(memnew(DebuggerEditorPlugin(this, debug_menu)));
@@ -6784,11 +6783,6 @@ EditorNode::EditorNode() {
 	} else {
 		WARN_PRINT("Asset Library not available, as it requires SSL to work.");
 	}
-
-	//add interface before adding plugins
-
-	editor_interface = memnew(EditorInterface);
-	add_child(editor_interface);
 
 	//more visually meaningful to have this later
 	raise_bottom_panel_item(AnimationPlayerEditor::singleton);
