+++ editor/property_selector.h
@@ -35,6 +35,8 @@
 #include "editor_help.h"
 #include "scene/gui/rich_text_label.h"
 
+class EditorHelpBit;
+
 class PropertySelector : public ConfirmationDialog {
 	GDCLASS(PropertySelector, ConfirmationDialog);
